// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  articles: typeof articles;
  auth: typeof auth;
  behavioral: typeof behavioral;
  contentRecommendations: typeof contentRecommendations;
  conversations: typeof conversations;
  crons: typeof crons;
  financialProducts: typeof financialProducts;
  financialProfiles: typeof financialProfiles;
  http: typeof http;
  messages: typeof messages;
  notifications: typeof notifications;
  profiles: typeof profiles;
  providerConnections: typeof providerConnections;
  recommendations: typeof recommendations;
  rss: typeof rss;
  userApplications: typeof userApplications;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;